refactor: split routes_evaluation.py into helpers and routes modules
Decompose the 134-line evaluation routes module into _evaluation_helpers
(validation logic, 78 lines) and _evaluation_routes (route registrations,
73 lines). The original routes_evaluation.py becomes a re-export hub
preserving backward compatibility.
